Update V1.3
Fixed the bug "index out of range". The error was caused due to the first (oldest) review not having a rating, this threw off the printing to the data file. Added/updated some comments for readability. Checked to make sure the 5,4,3,2,1 rating score bug was fixed: confirmed it is working correctly now. Checked to make sure the hidden paragraph bug was fixed: confirmed it is working correctly now.